[schema] Rename align.circuit to align.schema
