+++ align/schema/README.md
@@ -1,15 +1,15 @@
 # Introduction
 
-The align.circuit module attempts to formalize specification & communication of schematic and layout information used in various stages of the ALIGN toolkit.
+The align.schema module attempts to formalize specification & communication of schematic and layout information used in various stages of the ALIGN toolkit.
 
-While the classes and methods implemented in this module can also be used as a Domain Specific Language (DSL) for circuit representation, this is not our target goal at the current point in time.
+While the classes and methods implemented in this module can also be used as a Domain Specific Language (DSL) for analog circuit representation, this is not our target goal at the current point in time.
 
 ## Data representation
-Almost all the attributes for our API can be easily interpreted by simply viewing the type hints in our class definitions. This document instead focuses on how these classes may be used using concrete examples. Please view tests/circuit for additional examples.
+Almost all the attributes for our API can be easily interpreted by simply viewing the type hints in our class definitions. This document instead focuses on how these classes may be used using concrete examples. Please view tests/schema for additional examples.
 
 A couple of important things to note:
 1. All the classes do a certain level of data validation. Some of the classes might additionally do a certain level of casting. Parameter values are a unique example of this. Whether you provide int, float or string, the values will get converted to python strings for internal storage. Remember the type hings in our implementation refer to what the programmer should expect the type to be after an object has already been loaded in. It says nothing about the user input.
-2. Nearly all attributes get converted to upper-case internally. The simple reason behind this is that SPICE is case-insensitive. The most notable exception to this rule are the commands under align.circuit.constraints which are parsed as python expressions and are hence case sensitive.
+2. Nearly all attributes get converted to upper-case internally. The simple reason behind this is that SPICE is case-insensitive. The most notable exception to this rule are the commands under align.schema.constraints which are parsed as python expressions and are hence case sensitive.
 
 With that said, let us get started!
 
@@ -19,7 +19,7 @@ There are essentially two types of SPICE models:
 1. Elementary models (RES, CAP, IND, PMOS, NMOS etc.) which come pre-defined by a given simulator
 2. Derived models (.MODEL statements in SPICE) which are usually declared by the PDK
 
-Both of these are handled in align.circuit using a single Model class:
+Both of these are handled in align.schema using a single Model class:
 ```python
 class Model(schema.BaseModel):
     name : str                 # Model Name
@@ -55,7 +55,7 @@ The above statement is sort of equivalent to the SPICE statement:
 ```spice
 .MODEL NEWMOS TESTMOS PARAM3=3
 ```
-However, since parameters hold a lot of importance in the ALIGN flow, we find it easier to copy over parameter values to the inherited model instead of just providing a pointer. So the above model if dumped out from align.circuit.Model will look something like:
+However, since parameters hold a lot of importance in the ALIGN flow, we find it easier to copy over parameter values to the inherited model instead of just providing a pointer. So the above model if dumped out from align.schema.Model will look something like:
 ```spice
 .MODEL NEWMOS TESTMOS PARAM1=1.0 PARAM2=2 PARAM3=3
 ```
