+++ tests/schema/BUFFER_VCM_FINAL4.sp
